shared/pyproject.toml
[build-system]
requires = ["setuptools>=64.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "shared"
version = "0.1.2"
authors = [
    {name = "juandisay", email = "info@juandisay.org"},
]
description = "A shared module"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]

dependencies = [
    "requests>=2.25.0",
    "urllib3>=1.26.0",
    "python-dotenv>=1.0.1",
    "ipython>=8.12.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=20.8b1",
    "isort>=5.0",
    "mypy>=0.800",
]

[project.urls]
Homepage = "https://github.com/juandisay"


[tool.setuptools]
packages = ["shared"]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
sohoo = ["py.typed", "*.pyi", "**/*.pyi"]